refactor(breaking): Make WolframApp.build_code into Option<u32> (was u32)
Not all Wolfram apps make the build code easily discoverable without launching
a WL session, which isn't fast or reliable enough for first-pass discovery of
apps.

This is particulary hard on Linux.
